RAG PDF Agent/requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-community
python-dotenv
pypdf
